time_epochms return int64_t

Undo locker-priority changes - had dubious gains anyway


Undo verify: CAS64 is out-of-scope


Fix accidental typo

+++ bbinc/averager.h
@@ -21,13 +21,13 @@
 
 struct averager;
 struct averager *averager_new(int limit, int maxpoints);
-void averager_add(struct averager *avg, int value, int now);
+void averager_add(struct averager *avg, int value, int64_t now);
 double averager_avg(struct averager *avg);
 int averager_max(struct averager *avg);
 int averager_min(struct averager *avg);
 void averager_destroy(struct averager *avg);
 int averager_depth(struct averager *avg);
-void averager_purge_old(struct averager *avg, int now);
+void averager_purge_old(struct averager *avg, int64_t now);
 void averager_clear(struct averager *avg);
 
 struct point {
