+++ bbinc/epochlib.h
@@ -29,7 +29,7 @@ void time_epoch_(int32_t *epoch);
 int comdb2_time_epoch(void);
 
 /* RETURNS # MILLISECONDS SINCE STARTUP */
-int comdb2_time_epochms(void);
+int64_t comdb2_time_epochms(void);
 
 /* Microseconds since start of epoch */
 int64_t comdb2_time_epochus(void);
