+++ db/sql.h
@@ -859,7 +859,7 @@ struct sqlclntstate {
     int64_t num_resets;
     time_t connect_time;
     time_t last_reset_time;
-    int state_start_time;
+    int64_t state_start_time;
     enum connection_state state;
     pthread_mutex_t state_lk;
     /* The node doesn't change.  The pid does as connections get donated.  We
@@ -1091,9 +1091,9 @@ struct sql_thread {
     LINKC_T(struct sql_thread) lnk;
     pthread_mutex_t lk;
     struct Btree *bt, *bttmp;
-    int startms;
+    int64_t startms;
     int prepms;
-    int stime;
+    int64_t stime;
     int nmove;
     int nfind;
     int nwrite;
